june_2024_requirements.txt
absl-py==2.1.0
aiobotocore==2.5.4
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
anndata==0.10.5.post1
annoy==1.17.3
array_api_compat==1.7.1
asciitree==0.3.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Beaker==1.13.0
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.4.1
botocore==1.31.17
cachetools==5.3.3
celltypist==1.6.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorcet==3.1.0
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
dask==2024.6.2
dask-expr==1.1.5
dask-image==2024.5.3
datashader==0.16.2
debugpy==1.8.1
decorator==5.1.1
distributed==2024.6.2
dm-tree==0.1.8
dnspython==2.6.1
docrep==0.3.2
docutils==0.21.2
et-xmlfile==1.1.0
etils==1.7.0
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.19
fbpca==1.0
filelock==3.15.3
fiona==1.9.6
flatbuffers==24.3.25
flax==0.8.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2023.6.0
gast==0.4.0
gdown==5.2.0
geopandas==0.14.4
geosketch==1.2
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.64.1
gseapy==1.1.3
h11==0.14.0
h5py==3.11.0
hilearn==0.2.2
holoviews==1.19.0
hupper==1.12.1
idna==3.7
igraph==0.11.5
imageio==2.34.1
importlib_metadata==7.2.0
importlib_resources==6.4.0
inflect==7.3.0
intel-cmplr-lib-ur==2024.2.0
intel-openmp==2024.2.0
intervaltree==3.1.0
ipykernel==6.29.4
ipython==8.25.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
legacy-api-wrap==1.4
leidenalg==0.10.2
libclang==18.1.1
lightning==2.1.4
lightning-utilities==0.11.2
linkify-it-py==2.0.3
llvmlite==0.41.1
locket==1.0.0
loguru==0.7.2
louvain==0.8.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
matplotlib-scalebar==0.8.1
mdit-py-plugins==0.4.1
mdurl==0.1.2
mkl==2024.2.0
mkl-service==2.4.1
ml-collections==0.1.1
ml-dtypes==0.4.0
mongokit==0.9.1.1
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
mudata==0.2.3
multidict==6.0.5
multipledispatch==1.0.0
multiscale_spatial_image==0.11.2
muon==0.1.6
NaiveDE==1.2.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.3
newick==1.0.0
numba==0.58.0
numcodecs==0.12.1
numpy==1.25.2
numpyro==0.15.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ome-zarr==0.9.0
omnipath==1.0.8
opencv-python==4.10.0.84
openpyxl==3.1.4
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.20
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
panel==1.4.4
param==2.1.0
parso==0.8.4
partd==1.4.2
PasteDeploy==3.1.0
patsy==0.5.6
pexpect==4.9.0
piano==0.3
pillow==10.3.0
PIMS==0.7
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.19.6
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyamg==5.1.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyct==0.5.0
pygeos==0.14
Pygments==2.18.0
pymongo==4.7.3
pynndescent==0.5.13
pyparsing==3.1.2
pyproj==3.6.1
pyramid==2.0.2
pyramid-mako==1.1.0
pyramid_beaker==0.9
pyramid_debugtoolbar==4.12.1
pyro-api==0.1.2
pyro-ppl==1.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-igraph==0.11.5
pytorch-lightning==2.3.0
pytz==2024.1
pyviz_comms==3.0.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3fs==2023.6.0
scanorama==1.7.4
scanpy==1.9.8
scArches==0.6.1
scbsp==0.2.7
scHPL==1.0.5
scikit-image==0.21.0
scikit-learn==1.5.0
scikit-misc==0.3.1
scipy==1.11.4
scmags @ file:///mnt/work/RO_src/venv/stnav_venv/scmags
scvi-tools==1.1.2
seaborn==0.13.2
selenium==4.22.0
session-info==1.0.0
shapely==2.0.4
simplejson==3.19.2
six==1.16.0
slicerator==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpaGCN==1.2.7
sparse_dot_mkl==0.9.3
spatial_image==0.3.0
spatialdata==0.0.15
SpatialDE==1.1.3
SpatialDM==0.2.0
squidpy==1.5.0
stack-data==0.6.3
statsmodels==0.14.1
stdlib-list==0.10.0
stlearn==0.4.12
sympy==1.12.1
tacco==0.3.0
tbb==2021.13.0
tblib==3.0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-probability==0.14.0
tensorstore==0.1.62
termcolor==2.4.0
texttable==1.7.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.5.0
tifffile==2024.6.18
toolz==0.12.1
torch==2.3.1
torchmetrics==1.4.0.post0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
translationstring==1.4
trio==0.25.1
trio-websocket==0.11.1
triton==2.3.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
umap-learn==0.5.6
urllib3==1.26.19
validators==0.28.3
venusian==3.1.0
waitress==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
WebHelpers==1.3
WebOb==1.8.7
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
xarray==2024.6.0
xarray-dataclasses==1.8.0
xarray-datatree==0.0.14
xarray-schema==0.0.3
xarray-spatial==0.4.0
xyzservices==2024.6.0
yarl==1.9.4
zarr==2.18.2
zict==3.0.0
zipp==3.19.2
zope.deprecation==5.0
zope.interface==6.4.post2
